Move types to models to avoud circular dep
